11 try picorv32 (add half-UART in systemverilog)
This adds a couple systemverilog files for the TX side of a UART.
It sends "T" once per second at 19200 baud on the OrangeCrab's TX
pin as well as the red LED channel.

The main point of this was to have a module that reads data from
a bus when given a write strobe, and then uses that data to do
something observable. Hopefully, I should be able to plumb this up
to one of the PicoRV32's data bus options.